requirements.txt
Flask==2.1.1
Pillow==9.0.1
stegano==0.1.2
